1.升级剪辑SDK版本 2.新增部分demo功能 3.修改部分bug

+++ editor/src/main/java/com/quvideo/application/widget/seekbar/DoubleSeekbar.java
@@ -27,7 +27,7 @@ public class DoubleSeekbar extends View {
   private int minProgress;
   private int maxProgress = 100;
   private int minRange = 0;
-  private int count = maxProgress - minProgress;
+  private long count = maxProgress - minProgress;
 
   /** 第一格进度 */
   private int firstProgress;
@@ -90,7 +90,7 @@ public class DoubleSeekbar extends View {
   }
 
   public void setProgressRange(int start, int end, int minRange) {
-    if (start > end || end - start < minRange) {
+    if (start > end || ((long) end - start) < minRange) {
       return;
     }
     this.minProgress = start;
